Added array instruction counter
Implements a wrapper for an integer array.
It presents some helpful methods for getting or setting array values,
while counting the number of instruction used.
Other methods have also been added, such as a a way to retrieve the
minimum, or to check if the array is sorted. Please note that using
these methods will add all read and write to the total number of
instruction used.